requirements.txt
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cssselect2==0.6.0
defusedxml==0.7.1
Django==4.0
django-colorfield==0.7.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-base64==2.0
drf-extra-fields==3.4.0
flake8==4.0.1
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
fonttools==4.33.3
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==1.7.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
pep8-naming==0.13.0
Pillow==9.1.1
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycparser==2.21
pydyf==0.2.0
pyflakes==2.4.0
PyJWT==2.4.0
pyPdf==1.13
pyphen==0.12.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
reportlab==3.6.10
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
testfixtures==6.18.5
tinycss2==1.1.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
webcolors==1.12
webencodings==0.5.1
zipp==3.8.0
zopfli==0.2.1
